02/04.agent_with_multiple_tools/pyproject.toml
[project]
name = "agent-with-multiple-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
